Change links to relative css to google cdn
This 'fix' is here because in current setup, are having difficulty
compiling sass and properly importing the default themes below.  This is
because in development, we're using tutor, which does not store these
themes in the same manner as in production.  This is a quick fix, and if
can get sass working smoothly, should migrate to sass.